DRTII-232 Move test route to productions code
